src/elements/other/__tests__/block.test.tsx
import Enzyme from "enzyme";
import React from "react";

import { Block } from "../block";

import { hasProperties, testGenericPropTypes } from "@/__tests__/testing";

describe("Block component", () => {
  hasProperties(Block, {
    defaultProps: { as: "div" },
  });

  it("should render as the default element", () => {
    const wrapper = Enzyme.shallow(<Block />);
    expect(wrapper.is("div")).toBe(true);
  });

  it("should render as a custom component", () => {
    const as = "span";
    const wrapper = Enzyme.shallow(<Block as={as} />);
    expect(wrapper.is(as)).toBe(true);
  });

  it("should forward ref", () => {
    const ref = React.createRef<HTMLDivElement>();
    // Enzyme owns outer ref: https://github.com/airbnb/enzyme/issues/1852
    const wrapper = Enzyme.mount(
      <div>
        <Block ref={ref} />
      </div>,
    );
    try {
      expect(ref.current).toBe(wrapper.find(".block").instance());
    } finally {
      wrapper.unmount();
    }
  });

  it("should have bulma className", () => {
    const wrapper = Enzyme.shallow(<Block />);
    expect(wrapper.hasClass("block")).toBe(true);
  });

  it("should preserve custom className", () => {
    const className = "foo";
    const wrapper = Enzyme.shallow(<Block className={className} />);
    expect(wrapper.hasClass(className)).toBe(true);
  });

  describe("propTypes", () => {
    const { propTypes } = Block;
    testGenericPropTypes(propTypes);
  });
});
